pyproject.toml
[project]
name = "data-for-ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "huggingface-hub>=1.6.0",
    "kaggle>=2.0.0",
    "numpy>=2.4.3",
    "python-dotenv>=1.2.2",
    "rank-bm25>=0.2.2",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.2.3",
    "torch>=2.10.0",
]
